Move 3 functions to depmap_preprocessing.py
